src/test/java/ru/yandex/practicum/filmorate/dao/MpaDaoTest.java
package ru.yandex.practicum.filmorate.dao;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;

import static org.junit.jupiter.api.Assertions.assertTrue;

@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
class MpaDaoTest {
    private final MpaDao mpaDao;

    @Test
    void getMpaById() {
        assertTrue(mpaDao.getMpaById(1).getName().equals("G"));
    }

    @Test
    void getAllMpa() {
        assertTrue(mpaDao.getAllMpa().size() == 5);
    }
}